requirements-dev.txt
### Runtime dependencies

-r requirements.txt


### Development dependencies

# Debug toolbar
django-debug-toolbar

# Code formatting
black

# Code linting
flake8

# Test runner
pytest
pytest-cov
pytest-django
tox

# Dependency management tools
pip-tools
